+++ src/main/java/com/orvillex/bortus/utils/SecurityUtils.java
@@ -1,44 +1,28 @@
-/*
- *  Copyright 2019-2020 Zheng Jie
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package me.zhengjie.utils;
+package com.orvillex.bortus.utils;
 
 import cn.hutool.json.JSONArray;
 import cn.hutool.json.JSONObject;
 import cn.hutool.json.JSONUtil;
-import lombok.extern.slf4j.Slf4j;
-import me.zhengjie.exception.BadRequestException;
-import me.zhengjie.utils.enums.DataScopeEnum;
+import com.orvillex.bortus.exception.BadRequestException;
+import com.orvillex.bortus.handler.SpringContextHolder;
+import com.orvillex.bortus.enums.DataScopeType;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import java.util.List;
+import java.util.*;
+
 
 /**
- * 获取当前登录的用户
- * @author Zheng Jie
- * @date 2019-01-17
+ * 用户相关工具类
+ * @author y-z-f
+ * @version 0.1
  */
-@Slf4j
 public class SecurityUtils {
-
+    
     /**
      * 获取当前登录的用户
-     * @return UserDetails
      */
     public static UserDetails getCurrentUser() {
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -54,9 +38,7 @@ public class SecurityUtils {
     }
 
     /**
-     * 获取系统用户名称
-     *
-     * @return 系统用户名称
+     * 获取当前登录用户名
      */
     public static String getCurrentUsername() {
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -68,33 +50,30 @@ public class SecurityUtils {
     }
 
     /**
-     * 获取系统用户ID
-     * @return 系统用户ID
+     * 获取当前登录用户编号
      */
     public static Long getCurrentUserId() {
         UserDetails userDetails = getCurrentUser();
-        return new JSONObject(new JSONObject(userDetails).get("user")).get("id", Long.class);
+        return new JSONObject(new JSONObject(userDetails).get("user")).getLong("id");
     }
 
     /**
-     * 获取当前用户的数据权限
-     * @return /
+     * 获取当前登录用户数据权限
      */
-    public static List<Long> getCurrentUserDataScope(){
+    public static List<Long> getCurrentUserDataScope() {
         UserDetails userDetails = getCurrentUser();
         JSONArray array = JSONUtil.parseArray(new JSONObject(userDetails).get("dataScopes"));
-        return JSONUtil.toList(array,Long.class);
+        return JSONUtil.toList(array, Long.class);
     }
 
     /**
      * 获取数据权限级别
-     * @return 级别
      */
     public static String getDataScopeType() {
         List<Long> dataScopes = getCurrentUserDataScope();
-        if(dataScopes.size() != 0){
+        if (dataScopes.size() != 0) {
             return "";
         }
-        return DataScopeEnum.ALL.getValue();
+        return DataScopeType.ALL.getValue();
     }
 }