+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 修订记录  
 
-## 删除内容  
+## 相关内容  
 
 1. 移除OSHI机器性能信息采集相关实现，对应接口`/api/monitor/query`和权限为`monitor:list`;  
-2. 
+2. `CodeEnum`枚举调整到`CacheKey`中使用;  
+3.